GEODE-9291 sorted set benchmarks - rename RedisHash to RedisSplitKey
