+++ geode-benchmarks/src/main/java/org/apache/geode/benchmark/redis/tasks/PrePopulateRedisSortedSet.java
@@ -18,8 +18,8 @@
 package org.apache.geode.benchmark.redis.tasks;
 
 import static java.lang.String.valueOf;
-import static org.apache.geode.benchmark.redis.tasks.RedisSubpart.toPart;
-import static org.apache.geode.benchmark.redis.tasks.RedisSubpart.toKey;
+import static org.apache.geode.benchmark.redis.tasks.RedisSplitKey.toPart;
+import static org.apache.geode.benchmark.redis.tasks.RedisSplitKey.toKey;
 
 import org.apache.geode.benchmark.LongRange;
