+++ geode-benchmarks/src/main/java/org/apache/geode/benchmark/redis/tasks/RedisSplitKey.java
@@ -16,17 +16,17 @@
 package org.apache.geode.benchmark.redis.tasks;
 
 /**
- * Split keyspace into Redis hash key and field parts.
+ * Split keyspace into a sublevel of keys.
  */
-public class RedisSubpart {
+public class RedisSplitKey {
 
-  public static long NUM_SUBPARTS_PER_KEY = 1000;
+  public static long NUM_PARTS_PER_KEY = 1000;
 
   public static long toKey(final long key) {
-    return key / NUM_SUBPARTS_PER_KEY;
+    return key / NUM_PARTS_PER_KEY;
   }
 
   public static long toPart(final long key) {
-    return key % NUM_SUBPARTS_PER_KEY;
+    return key % NUM_PARTS_PER_KEY;
   }
 }
